Each picture only shows major differences in one color. The pictures were created through the nasty for loop and by changing the x's of the (x vals: 0,1,2) if i % 3 == x and the order of them. Why would that change the color that has the details in it?
:::::::PROBLEM FOUND AND FIXED:::::::

In the current code there are (naught, one, two) _count
variables. In the previous code it was just one variable
count. This variable was intended to be used in the
update_pixel function to update which row the pixel was
taken from. (You want to take the pixel from one row three
times to get the proper picture). Count would reset each
time it hit nine, as would row. The problem was that count
would only hit nine during one color, so only one color
would change its row. The other colors would be stuck on
the same row forever. The fix was simple, and easy.
Use three different variables, duhhh. This now works.

:::::::::FUTURE DEVELOPMENT::::::
-find the colors of perlers available, create a palette
 out of them.

-implement a dithering function with the perler.

-clean up this awful code.