BUG FIX function slash line 67
old code:
if num_shades < 1:

new code:
if num_shades > 1:

this line of code checks for dividing by zero, and values that
don't make sense and prevents those problems.
The old code made it so this function would only try to run
if you provided a number that didn't make sense.